Refactor monotonic and realtime clocks
Now we compute monotonic time using the existing "seconds" variable and
a new "monoticks" variable. The seconds variable has an arbitrary
reference point (though it should be fixed across system reboots). This
ensures that the monotonic clock is long-term stable, because it was
previously incremented inside the 256Hz timer interrupt which was prone
to losing ticks.

We also compute the realtime clock from the computed monotonic time and
a new "realtime_offset" variable. The "realtime_offset" variable itself
is adjusted (either with a jump or gradually each tick) in order to
adjust the realtime clock.

This change simplifies how the monotonic and realtime clocks are managed
and slightly reduces the 256Hz timer overhead.

Also adjust the realtime offset by the time adjustment set by adjtime()
when the calculator is off (without going backwards in time); processes
are halted while the calculator is off, so no process will notice the
sudden time change. I'm not sure this is necessarily a good idea anyway
as the user may be surprised by this behavior. A better idea may be to
jump the realtime offset by only a fraction of the time the calculator
is off (e.g. 500 ppm, to match the regular adjustment rate), so that the
adjtime() adjustment continues even while the calculator is off.